+++ junit/src/test/java/cucumber/runtime/junit/JUnitReporterTest.java
@@ -2,7 +2,10 @@ package cucumber.runtime.junit;
 
 import cucumber.api.PendingException;
 import cucumber.api.Result;
+import cucumber.api.TestStep;
 import cucumber.runner.EventBus;
+import cucumber.runtime.junit.JUnitReporter.EachTestNotifier;
+import cucumber.runtime.junit.JUnitReporter.NoTestNotifier;
 import cucumber.runtime.junit.PickleRunners.PickleRunner;
 import gherkin.pickles.PickleStep;
 import org.junit.AssumptionViolatedException;
@@ -11,19 +14,20 @@ import org.junit.runner.Description;
 import org.junit.runner.notification.Failure;
 import org.junit.runner.notification.RunNotifier;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Matchers;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static java.util.Arrays.asList;
+
 
 public class JUnitReporterTest {
 
@@ -31,35 +35,77 @@ public class JUnitReporterTest {
     private RunNotifier runNotifier;
 
     @Test
-    public void match_allow_started_ignored() {
-        createAllowStartedIgnoredReporter();
-        PickleStep runnerStep = mockStep();
-        Description runnerStepDescription = stepDescription(runnerStep);
-        PickleRunner pickleRunner = mockPickleRunner(runnerSteps(runnerStep));
-        when(pickleRunner.describeChild(runnerStep)).thenReturn(runnerStepDescription);
+    public void test_case_started_fires_test_started_for_pickle() {
+        createNonStrictReporter();
+        PickleRunner pickleRunner = mockPickleRunner(Collections.<PickleStep>emptyList());
         runNotifier = mock(RunNotifier.class);
-
         jUnitReporter.startExecutionUnit(pickleRunner, runNotifier);
-        jUnitReporter.handleStepStarted(runnerStep);
+
+        jUnitReporter.handleTestCaseStarted();
 
         verify(runNotifier).fireTestStarted(pickleRunner.getDescription());
-        verify(runNotifier).fireTestStarted(runnerStepDescription);
     }
 
     @Test
-    public void result_with_error() {
+    public void test_step_started_fires_test_started_for_step() {
         createNonStrictReporter();
-        Result result = mock(Result.class);
-        Throwable exception = mock(Throwable.class);
-        when(result.getError()).thenReturn(exception);
+        PickleStep runnerStep = mockStep();
+        PickleRunner pickleRunner = mockPickleRunner(runnerSteps(runnerStep));
+        runNotifier = mock(RunNotifier.class);
+        jUnitReporter.startExecutionUnit(pickleRunner, runNotifier);
 
+        jUnitReporter.handleStepStarted(runnerStep);
+
+        verify(runNotifier).fireTestStarted(pickleRunner.describeChild(runnerStep));
+    }
+
+    @Test
+    public void test_step_finished_fires_only_test_finished_for_passed_step() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
         Description description = mock(Description.class);
-        createRunNotifier(description);
+        setUpStepNotifierAndStepErrors(description);
+        Result result = mockResult(Result.Type.PASSED);
 
-        jUnitReporter.handleStepResult(result);
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        verify(runNotifier).fireTestFinished(description);
+    }
+
+    @Test
+    public void test_step_finished_fires_assumption_failed_and_test_finished_for_skipped_step() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Result result = mockResult(Result.Type.SKIPPED);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
 
         ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
-        verify(runNotifier).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertTrue(failure.getException() instanceof SkippedThrowable);
+        assertEquals("This step is skipped", failure.getException().getMessage());
+    }
+
+    @Test
+    public void test_step_finished_fires_assumption_failed_and_test_finished_for_skipped_step_with_assumption_violated() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Throwable exception = new AssumptionViolatedException("Oops");
+        Result result = mockResult(Result.Type.SKIPPED, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
 
         Failure failure = failureArgumentCaptor.getValue();
         assertEquals(description, failure.getDescription());
@@ -67,241 +113,349 @@ public class JUnitReporterTest {
     }
 
     @Test
-    public void result_with_assumption_violated() {
-        createStrictReporter();
+    public void test_step_finished_adds_no_step_exeption_for_skipped_step_without_exception() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        setUpStepNotifierAndStepErrors();
         Result result = mockResult(Result.Type.SKIPPED);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        assertTrue(jUnitReporter.stepErrors.isEmpty());
+    }
+
+    @Test
+    public void test_step_finished_adds_the_step_exeption_for_skipped_step_with_assumption_violated() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        setUpStepNotifierAndStepErrors();
         Throwable exception = new AssumptionViolatedException("Oops");
+        Result result = mockResult(Result.Type.SKIPPED, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        assertEquals(asList(exception), jUnitReporter.stepErrors);
+    }
+
+    @Test
+    public void test_step_finished_fires_assumption_failed_and_test_finished_for_pending_step_in_non_strict_mode() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Throwable exception = new PendingException();
+        Result result = mockResult(Result.Type.PENDING, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertEquals(exception, failure.getException());
+    }
+
+    @Test
+    public void test_step_finished_fires_assumption_failed_and_test_finished_for_pending_step_in_strict_mode() {
+        createStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Throwable exception = new PendingException();
+        Result result = mockResult(Result.Type.PENDING, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertEquals(exception, failure.getException());
+    }
+
+    @Test
+    public void test_step_finished_adds_the_step_exeption_for_pending_steps() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        setUpStepNotifierAndStepErrors();
+        Throwable exception = new PendingException();
+        Result result = mockResult(Result.Type.PENDING, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        assertEquals(asList(exception), jUnitReporter.stepErrors);
+    }
+
+    @Test
+    public void test_step_finished_fires_assumption_failed_and_test_finished_for_undefined_step_in_non_strict_mode() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Result result = mockResult(Result.Type.UNDEFINED);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertTrue(failure.getException() instanceof UndefinedThrowable);
+        assertEquals("This step is undefined", failure.getException().getMessage());
+    }
+
+    @Test
+    public void test_step_finished_fires_failure_and_test_finished_for_undefined_step_in_strict_mode() {
+        createStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Result result = mockResult(Result.Type.UNDEFINED);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertTrue(failure.getException() instanceof UndefinedThrowable);
+        assertEquals("This step is undefined", failure.getException().getMessage());
+    }
+
+    @Test
+    public void test_step_finished_adds_a_step_exeption_for_undefined_steps() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        setUpStepNotifierAndStepErrors();
+        TestStep testStep = mockTestStep("XX");
+        Result result = mockResult(Result.Type.UNDEFINED);
+
+        jUnitReporter.handleStepResult(testStep, result);
+
+        assertFalse(jUnitReporter.stepErrors.isEmpty());
+        assertEquals("The step \"XX\" is undefined", jUnitReporter.stepErrors.get(0).getMessage());
+    }
+
+    @Test
+    public void test_step_finished_fires_failure_and_test_finished_for_failed_step() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        setUpStepNotifierAndStepErrors(description);
+        Throwable exception = mock(Throwable.class);
+        Result result = mockResult(Result.Type.FAILED, exception);
+
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertEquals(exception, failure.getException());
+    }
+
+    @Test
+    public void test_step_finished_adds_the_step_exeption_for_failed_steps() {
+        createNonStrictReporter();
+        createDefaultRunNotifier();
+        setUpStepNotifierAndStepErrors();
+        Throwable exception = new PendingException();
+        Result result = mockResult(Result.Type.FAILED, exception);
         when(result.getError()).thenReturn(exception);
 
-        PickleStep runnerStep = mockStep();
-        Description runnerStepDescription = stepDescription(runnerStep);
-        PickleRunner pickleRunner = mockPickleRunner(runnerSteps(runnerStep));
-        when(pickleRunner.describeChild(runnerStep)).thenReturn(runnerStepDescription);
-        Description pickleRunnerDescription = mock(Description.class);
-        when(pickleRunner.getDescription()).thenReturn(pickleRunnerDescription);
+        jUnitReporter.handleStepResult(mock(TestStep.class), result);
 
+        assertEquals(asList(exception), jUnitReporter.stepErrors);
+    }
 
+    @Test
+    public void test_case_finished_fires_only_test_finished_for_passed_step() {
+        createNonStrictReporter();
         Description description = mock(Description.class);
         createRunNotifier(description);
+        Result result = mockResult(Result.Type.PASSED);
 
-        jUnitReporter.startExecutionUnit(pickleRunner, runNotifier);
-        jUnitReporter.handleStepStarted(runnerStep);
-        jUnitReporter.handleStepResult(result);
+        jUnitReporter.handleTestCaseResult(result);
+
+        verify(runNotifier).fireTestFinished(description);
+    }
+
+    @Test
+    public void test_case_finished_fires_assumption_failed_and_test_finished_for_skipped_step() {
+        createNonStrictReporter();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        populateStepErrors(Collections.<Throwable>emptyList());
+        Result result = mockResult(Result.Type.SKIPPED);
+
+        jUnitReporter.handleTestCaseResult(result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        Failure failure = failureArgumentCaptor.getValue();
+        assertEquals(description, failure.getDescription());
+        assertTrue(failure.getException() instanceof SkippedThrowable);
+    }
+
+    @Test
+    public void test_case_finished_fires_assumption_failed_and_test_finished_for_skipped_step_with_assumption_violated() {
+        createNonStrictReporter();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(AssumptionViolatedException.class);
+        Throwable exception2 = mock(AssumptionViolatedException.class);
+        populateStepErrors(asList(exception1, exception2));
+        Result result = mockResult(Result.Type.SKIPPED);
+
+        jUnitReporter.handleTestCaseResult(result);
 
         ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
         verify(runNotifier, times(2)).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
 
-        List<Failure> failure = failureArgumentCaptor.getAllValues();
-        assertEquals(runnerStepDescription, failure.get(0).getDescription());
-        assertEquals(exception, failure.get(0).getException());
-
-        assertEquals(pickleRunnerDescription, failure.get(1).getDescription());
-        assertEquals(exception, failure.get(1).getException());
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
     }
 
     @Test
-    public void result_with_undefined_step_non_strict() {
+    public void test_case_finished_fires_assumption_failed_and_test_finished_for_pending_step_in_non_strict_mode() {
         createNonStrictReporter();
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-
-        jUnitReporter.handleStepResult(mockResult(Result.Type.UNDEFINED));
-
-        verify(stepNotifier, times(0)).fireTestStarted();
-        verify(stepNotifier, times(0)).fireTestFinished();
-        verify(stepNotifier, times(0)).addFailure(Matchers.<Throwable>any(Throwable.class));
-        verify(stepNotifier).fireTestIgnored();
-    }
-
-    @Test
-    public void result_with_undefined_step_strict() {
-        createStrictReporter();
-        createDefaultRunNotifier();
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-        JUnitReporter.EachTestNotifier pickleRunnerNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.pickleRunnerNotifier = pickleRunnerNotifier;
-
-        jUnitReporter.handleStepResult(mockResult(Result.Type.UNDEFINED));
-
-        verify(stepNotifier, times(1)).fireTestStarted();
-        verify(stepNotifier, times(1)).fireTestFinished();
-        verifyAddFailureWithPendingException(stepNotifier);
-        verifyAddFailureWithPendingException(pickleRunnerNotifier);
-        verify(stepNotifier, times(0)).fireTestIgnored();
-    }
-
-    private void verifyAddFailureWithPendingException(JUnitReporter.EachTestNotifier stepNotifier) {
-        verifyAddFailureWithException(PendingException.class, stepNotifier);
-    }
-
-
-    private void verifyAddFailureWithException(Class<?> exception, JUnitReporter.EachTestNotifier stepNotifier) {
-        ArgumentCaptor<Throwable> captor = ArgumentCaptor.forClass(Throwable.class);
-        verify(stepNotifier).addFailure(captor.capture());
-        Throwable error = captor.getValue();
-        assertTrue(exception.isInstance(error));
-    }
-
-    @Test
-    public void result_with_pending_step_non_strict() {
-        createNonStrictReporter();
-        Result result = mock(Result.class);
-        when(result.getError()).thenReturn(new PendingException());
-
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-
-        jUnitReporter.handleStepResult(result);
-
-        verify(stepNotifier, times(0)).fireTestStarted();
-        verify(stepNotifier, times(0)).fireTestFinished();
-        verify(stepNotifier, times(0)).addFailure(Matchers.<Throwable>any(Throwable.class));
-        verify(stepNotifier).fireTestIgnored();
-    }
-
-    @Test
-    public void result_with_pending_step_strict() {
-        createStrictReporter();
-        createDefaultRunNotifier();
-        Result result = mock(Result.class);
-        when(result.getError()).thenReturn(new PendingException());
-
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-        JUnitReporter.EachTestNotifier pickleRunnerNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.pickleRunnerNotifier = pickleRunnerNotifier;
-
-        jUnitReporter.handleStepResult(result);
-
-        verify(stepNotifier, times(1)).fireTestStarted();
-        verify(stepNotifier, times(1)).fireTestFinished();
-        verifyAddFailureWithPendingException(stepNotifier);
-        verifyAddFailureWithPendingException(pickleRunnerNotifier);
-        verify(stepNotifier, times(0)).fireTestIgnored();
-    }
-
-    @Test
-    public void result_without_error_non_strict() {
-        createNonStrictReporter();
-        Result result = mock(Result.class);
-
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-
-        jUnitReporter.handleStepResult(result);
-
-        verify(stepNotifier).fireTestStarted();
-        verify(stepNotifier).fireTestFinished();
-        verify(stepNotifier, times(0)).addFailure(Matchers.<Throwable>any(Throwable.class));
-        verify(stepNotifier, times(0)).fireTestIgnored();
-    }
-
-    @Test
-    public void result_without_error_strict() {
-        createStrictReporter();
-        Result result = mock(Result.class);
-
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-
-        jUnitReporter.handleStepResult(result);
-
-        verify(stepNotifier).fireTestStarted();
-        verify(stepNotifier).fireTestFinished();
-        verify(stepNotifier, times(0)).addFailure(Matchers.<Throwable>any(Throwable.class));
-        verify(stepNotifier, times(0)).fireTestIgnored();
-    }
-
-    @Test
-    public void result_without_error_allow_stared_ignored() {
-        createAllowStartedIgnoredReporter();
-        Result result = mock(Result.class);
-
-        JUnitReporter.EachTestNotifier stepNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.stepNotifier = stepNotifier;
-
-        jUnitReporter.handleStepResult(result);
-
-        verify(stepNotifier, times(0)).fireTestStarted();
-        verify(stepNotifier).fireTestFinished();
-        verify(stepNotifier, times(0)).addFailure(Matchers.<Throwable>any(Throwable.class));
-        verify(stepNotifier, times(0)).fireTestIgnored();
-    }
-
-    @Test
-    public void hook_with_pending_exception_strict() {
-        createStrictReporter();
-        createDefaultRunNotifier();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(Throwable.class);
+        Throwable exception2 = mock(Throwable.class);
+        populateStepErrors(asList(exception1, exception2));
         Result result = mockResult(Result.Type.PENDING);
-        when(result.getError()).thenReturn(new PendingException());
 
-        JUnitReporter.EachTestNotifier pickleRunnerNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.pickleRunnerNotifier = pickleRunnerNotifier;
+        jUnitReporter.handleTestCaseResult(result);
 
-        jUnitReporter.handleHookResult(result);
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier, times(2)).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
 
-        verifyAddFailureWithPendingException(pickleRunnerNotifier);
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
     }
 
     @Test
-    public void hook_with_pending_exception_non_strict() {
-        createNonStrictReporter();
-        createDefaultRunNotifier();
+    public void test_case_finished_fires_failure_and_test_finished_for_pending_step_in_strict_mode() {
+        createStrictReporter();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(Throwable.class);
+        Throwable exception2 = mock(Throwable.class);
+        populateStepErrors(asList(exception1, exception2));
         Result result = mockResult(Result.Type.PENDING);
-        when(result.getError()).thenReturn(new PendingException());
 
-        JUnitReporter.EachTestNotifier pickleRunnerNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.pickleRunnerNotifier = pickleRunnerNotifier;
+        jUnitReporter.handleTestCaseResult(result);
 
-        jUnitReporter.handleHookResult(result);
-        jUnitReporter.finishExecutionUnit();
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier, times(2)).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
 
-        verify(pickleRunnerNotifier).fireTestIgnored();
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
     }
 
     @Test
-    public void failed_step_and_after_with_pending_exception_non_strict() {
+    public void test_case_finished_fires_assumption_failed_and_test_finished_for_undefined_step_in_non_strict_mode() {
         createNonStrictReporter();
-        createDefaultRunNotifier();
-        Result stepResult = mock(Result.class);
-        Throwable exception = mock(Throwable.class);
-        when(stepResult.getError()).thenReturn(exception);
-        Result hookResult = mockResult(Result.Type.PENDING);
-        when(hookResult.getError()).thenReturn(new PendingException());
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(Throwable.class);
+        Throwable exception2 = mock(Throwable.class);
+        populateStepErrors(asList(exception1, exception2));
+        Result result = mockResult(Result.Type.UNDEFINED);
 
-        JUnitReporter.EachTestNotifier pickleRunnerNotifier = mock(JUnitReporter.EachTestNotifier.class);
-        jUnitReporter.pickleRunnerNotifier = pickleRunnerNotifier;
+        jUnitReporter.handleTestCaseResult(result);
 
-        jUnitReporter.handleStepResult(stepResult);
-        jUnitReporter.handleHookResult(hookResult);
-        jUnitReporter.finishExecutionUnit();
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier, times(2)).fireTestAssumptionFailed(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
 
-        verify(pickleRunnerNotifier, times(0)).fireTestIgnored();
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
     }
 
     @Test
-    public void creates_step_notifier_with_step_from_execution_unit_runner() throws Exception {
-        PickleStep runnerStep = mockStep("Step Name");
-        Description runnerStepDescription = stepDescription(runnerStep);
-        PickleRunner pickleRunner = mockPickleRunner(runnerSteps(runnerStep));
-        when(pickleRunner.describeChild(runnerStep)).thenReturn(runnerStepDescription);
-        RunNotifier notifier = mock(RunNotifier.class);
-        jUnitReporter = new JUnitReporter(mock(EventBus.class), false, new JUnitOptions(Collections.<String>emptyList()));
+    public void test_case_finished_fires_failure_and_test_finished_for_undefined_step_in_strict_mode() {
+        createStrictReporter();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(Throwable.class);
+        Throwable exception2 = mock(Throwable.class);
+        populateStepErrors(asList(exception1, exception2));
+        Result result = mockResult(Result.Type.UNDEFINED);
 
-        jUnitReporter.startExecutionUnit(pickleRunner, notifier);
-        jUnitReporter.handleStepStarted(runnerStep);
-        jUnitReporter.handleStepResult(mockResult());
+        jUnitReporter.handleTestCaseResult(result);
 
-        verify(notifier).fireTestFinished(runnerStepDescription);
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier, times(2)).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
     }
 
-    private Result mockResult() {
-        return mockResult(Result.Type.PASSED);
+    @Test
+    public void test_case_finished_fires_failure_and_test_finished_for_failed_step() {
+        createNonStrictReporter();
+        Description description = mock(Description.class);
+        createRunNotifier(description);
+        Throwable exception1 = mock(Throwable.class);
+        Throwable exception2 = mock(Throwable.class);
+        populateStepErrors(asList(exception1, exception2));
+        Result result = mockResult(Result.Type.FAILED);
+
+        jUnitReporter.handleTestCaseResult(result);
+
+        ArgumentCaptor<Failure> failureArgumentCaptor = ArgumentCaptor.forClass(Failure.class);
+        verify(runNotifier, times(2)).fireTestFailure(failureArgumentCaptor.capture());
+        verify(runNotifier).fireTestFinished(description);
+
+        List<Failure> failures = failureArgumentCaptor.getAllValues();
+        assertEquals(description, failures.get(0).getDescription());
+        assertEquals(exception1, failures.get(0).getException());
+        assertEquals(description, failures.get(1).getDescription());
+        assertEquals(exception2, failures.get(1).getException());
+    }
+
+    private Result mockResult(Result.Type status, Throwable exception) {
+        Result result = mockResult(status);
+        when(result.getError()).thenReturn(exception);
+        return result;
     }
 
     private Result mockResult(Result.Type status) {
         Result result = mock(Result.class);
+        when(result.getStatus()).thenReturn(status);
         for (Result.Type type : Result.Type.values()) {
             when(result.is(type)).thenReturn(type == status);
         }
@@ -311,6 +465,10 @@ public class JUnitReporterTest {
     private PickleRunner mockPickleRunner(List<PickleStep> runnerSteps) {
         PickleRunner pickleRunner = mock(PickleRunner.class);
         when(pickleRunner.getDescription()).thenReturn(mock(Description.class));
+        for (PickleStep runnerStep : runnerSteps) {
+            Description runnerStepDescription = stepDescription(runnerStep);
+            when(pickleRunner.describeChild(runnerStep)).thenReturn(runnerStepDescription);
+        }
         return pickleRunner;
     }
 
@@ -335,6 +493,12 @@ public class JUnitReporterTest {
         return step;
     }
 
+    private TestStep mockTestStep(String stepText) {
+        TestStep testStep = mock(TestStep.class);
+        when(testStep.getStepText()).thenReturn(stepText);
+        return testStep;
+    }
+
     private void createDefaultRunNotifier() {
         createRunNotifier(mock(Description.class));
     }
@@ -347,20 +511,31 @@ public class JUnitReporterTest {
     }
 
     private void createStrictReporter() {
-        createReporter(true, false);
+        createReporter(true);
     }
 
     private void createNonStrictReporter() {
-        createReporter(false, false);
+        createReporter(false);
     }
 
-    private void createAllowStartedIgnoredReporter() {
-        createReporter(false, true);
+    private void createReporter(boolean strict) {
+        jUnitReporter = new JUnitReporter(mock(EventBus.class), strict, new JUnitOptions(Collections.<String>emptyList()));
     }
 
-    private void createReporter(boolean strict, boolean allowStartedIgnored) {
-        String allowStartedIgnoredOption = allowStartedIgnored ? "--allow-started-ignored" : "--no-allow-started-ignored";
-        jUnitReporter = new JUnitReporter(mock(EventBus.class), strict, new JUnitOptions(asList(allowStartedIgnoredOption)));
+    private void setUpStepNotifierAndStepErrors(Description description) {
+        jUnitReporter.stepNotifier = new EachTestNotifier(runNotifier, description);
+        jUnitReporter.stepErrors = new ArrayList<Throwable>();
     }
 
+    private void setUpStepNotifierAndStepErrors() {
+        jUnitReporter.stepNotifier = new NoTestNotifier();
+        jUnitReporter.stepErrors = new ArrayList<Throwable>();
+    }
+
+    private void populateStepErrors(List<Throwable> exceptions) {
+        jUnitReporter.stepErrors = new ArrayList<Throwable>();
+        for (Throwable exception : exceptions) {
+            jUnitReporter.stepErrors.add(exception);
+        }
+    }
 }